Arreglado el editar un puesto y añadido que el admin pueda especificar el recinto asociado al crear y editar un puesto. Faltan validaciones

+++ src/main/java/org/springframework/samples/petclinic/web/PuestoController.java
@@ -1,7 +1,9 @@
 package org.springframework.samples.petclinic.web;
 
 import java.util.Collection;
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import javax.validation.Valid;
 
@@ -9,6 +11,7 @@ import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.samples.petclinic.model.Festival;
 import org.springframework.samples.petclinic.model.Puesto;
+import org.springframework.samples.petclinic.model.Recinto;
 import org.springframework.samples.petclinic.model.TipoPuesto;
 import org.springframework.samples.petclinic.model.TipoTamaño;
 import org.springframework.samples.petclinic.service.FestivalService;
@@ -17,9 +20,7 @@ import org.springframework.samples.petclinic.service.RecintoService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.InitBinder;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -51,11 +52,11 @@ public class PuestoController {
 	public Optional<Festival> findFestival(@PathVariable("festivalId") int festivalId) {
 		return this.festivalService.findFestivalById(festivalId);
 	}
-
-	@InitBinder("festival")
-	public void initFestivalBinder(WebDataBinder dataBinder) {
-		dataBinder.setDisallowedFields("id");
-	}
+//
+//	@InitBinder("festival")
+//	public void initFestivalBinder(WebDataBinder dataBinder) {
+//		dataBinder.setDisallowedFields("id");
+//	}
 
 	@GetMapping
 	public String listPuestos(@PathVariable("festivalId") int festivalId, ModelMap model) {
@@ -74,6 +75,12 @@ public class PuestoController {
 		return this.puestoService.findTiposTamaño();
 	}
 
+	@ModelAttribute("recintos")
+	public List<String> recintos(@PathVariable("festivalId") int festivalId) {
+		return recintoService.findAllRecintosByFestivalId(festivalId).stream().map(r -> r.getName())
+				.collect(Collectors.toList());
+	}
+
 	@GetMapping("/new")
 	public String initCreationNewPuesto(ModelMap model) {
 		Puesto puesto = new Puesto();
@@ -90,20 +97,29 @@ public class PuestoController {
 		} else {
 			TipoPuesto tipoPuesto = puestoService.findPuestoType(puesto.getTipoPuesto().getName());
 			TipoTamaño tipoTamanio = puestoService.findTamañoType(puesto.getTipoTamanio().getName());
+			Recinto recinto = recintoService.findRecintoByName(puesto.getRecinto().getName());
 			puesto.setFestival(festivalService.findFestivalById(festivalId).get());
 			puesto.setTipoPuesto(tipoPuesto);
 			puesto.setTipoTamanio(tipoTamanio);
-//			puesto.setRecinto(this.recintoService.findRecintoByName(puesto.getRecinto().getName()));
+			puesto.setRecinto(recinto);
 			puestoService.save(puesto);
+			model.addAttribute("message", "Puesto creado correctamente!");
 			return "redirect:/festivales/{festivalId}/puestos";
 		}
 	}
 
 	@GetMapping("/{id}/delete")
 	public String deletePuesto(@PathVariable("id") int id, @PathVariable("festivalId") int festivalId, ModelMap model) {
-		Optional<Puesto> puesto = puestoService.findById(id);
-		puestoService.delete(puesto.get());
-		return "redirect:/festivales/{festivalId}/puestos";
+		if (this.puestoService.findById(id).isPresent()) {
+			Optional<Puesto> puesto = puestoService.findById(id);
+			puestoService.delete(puesto.get());
+			model.addAttribute("message", "Puesto borrado correctamente!");
+			return "redirect:/festivales/{festivalId}/puestos";
+		} else {
+			model.addAttribute("message", "No se encuentra!");
+			return "redirect:/festivales/{festivalId}/puestos";
+		}
+
 	}
 
 	@GetMapping("/{id}/edit")
@@ -115,21 +131,20 @@ public class PuestoController {
 
 	@PostMapping("/{id}/edit")
 	public String processUpdatePuesto(@PathVariable("id") int id, @PathVariable("festivalId") int festivalId,
-			@Valid Puesto puesto, BindingResult binding, ModelMap model){
-	
+			@Valid Puesto puesto, BindingResult binding, ModelMap model) {
+		Puesto modifiedPuesto = puestoService.findById(id).orElse(null);
 		if (binding.hasErrors()) {
 			model.put("puesto", puesto);
 			return PUESTOS_FORM;
 
 		} else {
-			Puesto modifiedPuesto = puestoService.findById(id).orElse(null);
-			BeanUtils.copyProperties(puesto, modifiedPuesto, "id", "festival");
-			TipoPuesto tipoPuesto = puestoService.findPuestoType(modifiedPuesto.getTipoPuesto().getName());
-			TipoTamaño tipoTamanio = puestoService.findTamañoType(modifiedPuesto.getTipoTamanio().getName());
+			BeanUtils.copyProperties(puesto, modifiedPuesto, "id");
+			modifiedPuesto.setTipoPuesto(this.puestoService.findPuestoType(puesto.getTipoPuesto().getName()));
+			modifiedPuesto.setTipoTamanio(this.puestoService.findTamañoType(puesto.getTipoTamanio().getName()));
 			modifiedPuesto.setFestival(festivalService.findFestivalById(festivalId).get());
-			modifiedPuesto.setTipoPuesto(tipoPuesto);
-			modifiedPuesto.setTipoTamanio(tipoTamanio);
-			puestoService.save(modifiedPuesto);
+			modifiedPuesto.setRecinto(recintoService.findRecintoByName(puesto.getRecinto().getName()));
+			this.puestoService.save(modifiedPuesto);
+			model.addAttribute("message", "Puesto actualizado correctamente!");
 			return "redirect:/festivales/{festivalId}/puestos";
 
 		}
